video: tegra: typecast atomic64_t to long long
atomic64_read returns data of type atomic64_t, which used to be
of type 'long' until K4.14. Starting K5.9, it has been changed
to s64 (long long). Instead of differentiating between the return
data between kernel versions, typecasted the data returned from
atomic64_read to long long (the bigger type).

bug 200601926

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2361769
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/video/tegra/dc/dc.c
@@ -2832,12 +2832,12 @@ static int dbg_flip_stats_show(struct seq_file *m, void *unused)
 	if (WARN_ON(!dc || !dc->out))
 		return -EINVAL;
 
-	seq_printf(m, "Flips queued: %ld\n",
-		atomic64_read(&dc->flip_stats.flips_queued));
-	seq_printf(m, "Flips skipped: %ld\n",
-		atomic64_read(&dc->flip_stats.flips_skipped));
-	seq_printf(m, "Flips completed: %ld\n",
-		atomic64_read(&dc->flip_stats.flips_cmpltd));
+	seq_printf(m, "Flips queued: %lld\n",
+		(long long int)atomic64_read(&dc->flip_stats.flips_queued));
+	seq_printf(m, "Flips skipped: %lld\n",
+		(long long int)atomic64_read(&dc->flip_stats.flips_skipped));
+	seq_printf(m, "Flips completed: %lld\n",
+		(long long int)atomic64_read(&dc->flip_stats.flips_cmpltd));
 
 	return 0;
 }
